Small withAutoMarkdown refactor
